launch: Build run manifest and invoke the manifest
Refactor invoke_run() so that it takes a RunManifest argument instead
of a Run. The RunManifest is constructed from the Run but contains
all data that is necessary to launch the run. This way, we can store
the manifest to a file or transmit it to a remote server before
launching the run.