launch/common: Support 'stdout' key for experiments and enhance 'output' key
Specifying 'stdout' as value for the 'output' key will trigger a DeprecationWarning. It is possible to specify output extensions by adding the 'extensions' key to the 'output' dictionary. Furthermore, it is now possible to specify the stdout path extensions of experiments via the 'stdout' key, e.g. stdout: 'ext' creates stdout files with extension .ext. A new variable 'OUTPUT:<ext>', where <ext> is an extension specified in the 'extensions' key ('out' will automatically be added) is supported in the 'args' key of experiments.